+++ src/main/java/com/ericsson/bss/cassandra/ecaudit/auth/WhitelistOptionParser.java
@@ -17,7 +17,10 @@ package com.ericsson.bss.cassandra.ecaudit.auth;
 
 import java.util.Set;
 
+import com.google.common.collect.ImmutableSet;
+
 import org.apache.cassandra.auth.IResource;
+import org.apache.cassandra.auth.Permission;
 import org.apache.cassandra.exceptions.InvalidRequestException;
 
 class WhitelistOptionParser
@@ -43,16 +46,23 @@ class WhitelistOptionParser
         }
     }
 
-    String parseTargetOperation(String inputOption)
+    Set<Permission> parseTargetOperation(String inputOption, IResource resource)
     {
         String operationString = stripOptionPrefix(inputOption).toUpperCase();
 
         if (ALL_OPERATIONS.equals(operationString))
         {
-            return ALL_OPERATIONS;
+            return resource.applicablePermissions();
         }
 
-        throw new InvalidRequestException("Invalid whitelist option: " + inputOption);
+        try
+        {
+            return ImmutableSet.of(Permission.valueOf(operationString));
+        }
+        catch (IllegalArgumentException e)
+        {
+            throw new InvalidRequestException("Invalid whitelist option: " + e.getMessage());
+        }
     }
 
     private String stripOptionPrefix(String option)
@@ -60,15 +70,15 @@ class WhitelistOptionParser
         return option.replaceFirst(VALID_PREFIX, "");
     }
 
-    Set<IResource> parseResource(String resourceNameCsv)
+    IResource parseResource(String resourceName)
     {
         try
         {
-            return ResourceFactory.toResourceSet(resourceNameCsv);
+            return ResourceFactory.toResource(resourceName);
         }
         catch (IllegalArgumentException e)
         {
-            throw new InvalidRequestException(String.format("Unable to parse whitelisted resource(s) [%s]: %s", resourceNameCsv, e.getMessage()));
+            throw new InvalidRequestException(String.format("Unable to parse whitelisted resource [%s]: %s", resourceName, e.getMessage()));
         }
     }
 }