+++ src/main/java/com/ericsson/bss/cassandra/ecaudit/entry/AuditEntry.java
@@ -223,6 +223,5 @@ public class AuditEntry
         {
             return new AuditEntry(this);
         }
-
     }
 }